Make Tone Numerator and Denominator 64 bit integers (#60)
For, like, 53 wide pythagorean scales and crazy stuff
which gives you massive fractions

+++ include/Tunings.h
@@ -59,7 +59,7 @@ struct Tone
 
     Type type;
     double cents;
-    int ratio_d, ratio_n;
+    int64_t ratio_d, ratio_n;
     std::string stringRep;
     double floatValue; // cents / 1200 + 1.
